requirements.txt
newspaper3k
simplejson

